Remove lol defines

+++ include/libft/mat.h
@@ -22,9 +22,9 @@ typedef struct	s_complex
 	double		i;
 }				t_complex;
 
-extern int32_t	ft_32npow(intmax_t nb, uint8_t power);
-extern int64_t	ft_64npow(intmax_t nb, uint8_t power);
-extern size_t	ft_stnpow(intmax_t nb, uint8_t power);
+extern int32_t	ft_32npow(int32_t nb, uint8_t power);
+extern int64_t	ft_64npow(int64_t nb, uint8_t power);
+extern size_t	ft_stnpow(size_t nb, uint8_t power);
 extern size_t	ft_stnxpow2(size_t nb);
 extern double	ft_zpowi(t_complex z, uint8_t n);
 extern double	ft_zpowr(t_complex z, uint8_t n);
